Move CriteriaTable into separate component
